Add FormStores::SystemAdmin::SchoolFormStore
* Extract common login to Cacheable
