+++ app/services/form_store.rb
@@ -1,7 +1,8 @@
+
 # frozen_string_literal: true
 
 class FormStore
-  class InvalidKeyError < StandardError; end
+  include Cacheable
 
   # Will need to remove lead_school when lead_partner transfer is complete
   FORM_SECTION_KEYS = %i[
@@ -48,37 +49,4 @@ class FormStore
     placements
     itt_end_date
   ].freeze
-
-  class << self
-    def get(trainee_id, key)
-      value = redis.get(cache_key_for(trainee_id, key))
-      JSON.parse(value) if value.present?
-    end
-
-    def set(trainee_id, key, values)
-      raise(InvalidKeyError) unless FORM_SECTION_KEYS.include?(key)
-
-      redis.set(cache_key_for(trainee_id, key), values.to_json)
-
-      true
-    end
-
-    def clear_all(trainee_id)
-      FORM_SECTION_KEYS.each do |key|
-        redis.set(cache_key_for(trainee_id, key), nil)
-      end
-    end
-
-    def cache_key_for(trainee_id, key)
-      if ENV["TEST_ENV_NUMBER"].present?
-        "#{trainee_id}_#{key}_#{ENV['TEST_ENV_NUMBER']}"
-      else
-        "#{trainee_id}_#{key}"
-      end
-    end
-
-    def redis
-      RedisClient.current
-    end
-  end
 end
