+++ app/services/form_stores/system_admin/school_form_store.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module FormStores
+  module SystemAdmin
+    class SchoolFormStore
+      include Cacheable
+
+      FORM_SECTION_KEYS = %i[
+        school
+      ].freeze
+    end
+  end
+end
